refactor(studio): extract ledger support services
